move object pan tilt tracker
